rename AccountQueryContainer to ProfileQueryContainer
