+++ resources/assets/components/pages/ProfilePage/ProfilePage.js
@@ -1,11 +1,11 @@
 import React from 'react';
 
 import Enclosure from '../../Enclosure';
-import AccountQueryContainer from './AccountQueryContainer';
+import ProfileQueryContainer from './ProfileQueryContainer';
 
 const ProfilePage = props => (
   <Enclosure className="container margin-bottom-lg">
-    <AccountQueryContainer {...props} />
+    <ProfileQueryContainer {...props} />
   </Enclosure>
 );
